Extract CaptureEngine from CameraViewController (UI modernization PR 8)
Mechanical extraction, no behavior or threading changes. CaptureEngine
(non-UI, NSObject, the AVCapturePhotoCaptureDelegate) now owns the
capture session, device/lens/zoom discovery, capabilities gathering,
flash/torch intent, quality/format/aspect configuration, and still
capture + crop math.

CameraViewController (~1,670 → ~960 lines) keeps the preview layer,
overlays, permissions, lifecycle, and the recording/sample-buffer
pipeline (queued for a later PR), reaching the session through
engine-exposed properties. Engine-to-UI seams: onPicture (relayed to
the actor system by the VC), onStatusChanged, and
rotateOutputs(orientation:) — the VC still rotates the preview-layer
connection, preserving the original method's exact shape.

CropRectTests re-point to CaptureEngine.cropRect (assertions
unchanged); adds 4 CaptureEngineTests for session-free logic.

376/376 tests pass (372 existing + 4 new), verified in two independent
full runs.
